# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IBKR Control Center",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/setup/state": {
      "get": {
        "tags": [
          "setup"
        ],
        "summary": "Get State",
        "operationId": "get_state_api_setup_state_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WizardStateResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/setup/step1/save": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Step1 Save",
        "description": "Save creds (encrypt token). Does NOT call IBKR per spec D5.",
        "operationId": "step1_save_api_setup_step1_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlexCredentialsValidate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Step1 Save Api Setup Step1 Save Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/setup/step2/detect": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Step2 Detect",
        "description": "Fetch + parse + persist YTD. Returns detected accounts (sin F).\n\nRetry policy per spec D10: server-side retry on 1001 with backoff [5, 15, 30]s.",
        "operationId": "step2_detect_api_setup_step2_detect_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Step2DetectResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/setup/step2/detect_from_xml": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Step2 Detect From Xml",
        "description": "Fallback when IBKR is unreachable: parse uploaded XML and persist it as\na manual upload so step2/save can validate detected accounts. The persister\ndedups by SHA-256, so re-uploads of the same XML are idempotent.",
        "operationId": "step2_detect_from_xml_api_setup_step2_detect_from_xml_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_step2_detect_from_xml_api_setup_step2_detect_from_xml_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Step2DetectFromXmlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/setup/step2/save": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Step2 Save",
        "description": "Persist accounts + participations. Dispatches TRM backfill in background (per D6).",
        "operationId": "step2_save_api_setup_step2_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step2SaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Step2SaveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/setup/step3/upload": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Step3 Upload",
        "description": "Stash one XML in memory. Detect new accounts vs already-configured.",
        "operationId": "step3_upload_api_setup_step3_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_step3_upload_api_setup_step3_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Step3UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/setup/step3/save_new_accounts": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Step3 Save New Accounts",
        "description": "Persist accounts + participations for newly-detected IDs from Step 3 uploads.",
        "operationId": "step3_save_new_accounts_api_setup_step3_save_new_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step3SaveNewAccountsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Step3 Save New Accounts Api Setup Step3 Save New Accounts Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/setup/step3/commit": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Step3 Commit",
        "description": "Drain stash, persist all selected XMLs in one transaction.\n\nValidates first that every detected (non-shadow) account in each stashed\nXML is already configured (via step3/save_new_accounts or earlier steps).",
        "operationId": "step3_commit_api_setup_step3_commit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step3CommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Step3CommitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/setup/finish": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Finish",
        "description": "Mark setup_completed_at. Validates preconditions (creds + at least 1 participation + step3 marked).",
        "operationId": "finish_api_setup_finish_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Finish Api Setup Finish Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FlexCredentialsValidate": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 10,
            "title": "Token"
          },
          "query_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Query Id"
          }
        },
        "type": "object",
        "required": [
          "token",
          "query_id"
        ],
        "title": "FlexCredentialsValidate"
      },
...
      "WizardStateResponse": {
        "properties": {
          "step1_credentials": {
            "type": "boolean",
            "title": "Step1 Credentials"
          },
          "step2_accounts": {
            "type": "boolean",
            "title": "Step2 Accounts"
          },
          "step3_xmls": {
            "type": "boolean",
            "title": "Step3 Xmls"
          },
          "step3_n_xmls_uploaded": {
            "type": "integer",
            "title": "Step3 N Xmls Uploaded"
          },
          "setup_completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setup Completed At"
          },
          "detected_accounts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DetectedAccount"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Accounts"
          },
          "pending_stash_temp_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pending Stash Temp Ids"
          }
        },
        "type": "object",
        "required": [
          "step1_credentials",
          "step2_accounts",
          "step3_xmls",
          "step3_n_xmls_uploaded",
          "setup_completed_at"
        ],
        "title": "WizardStateResponse"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}